+++ event/event.h
@@ -25,11 +25,11 @@ class Event
 {
 public:
     Event();
-    Event(const uint16_t id, const uint32_t param = 0);
+    Event(const uint8_t id, const uint32_t param = 0);
 
     bool operator== (const Event &rhs) const;
 
-    const uint16_t id;
+    const uint8_t id;
     const uint32_t param;
     uint32_t delay;
     uint32_t posted;
